+++ gameserver/src/main/java/com/shnok/javaserver/dto/external/serverpackets/ApplyDamagePacket.java
@@ -4,12 +4,12 @@ import com.shnok.javaserver.dto.SendablePacket;
 import com.shnok.javaserver.enums.network.packettypes.external.ServerPacketType;
 
 public class ApplyDamagePacket extends SendablePacket {
-    public ApplyDamagePacket(int sender, int target, int damage, int newHp, boolean criticalHit) {
+    public ApplyDamagePacket(int sender, int target, int damage, float newHp, boolean criticalHit) {
         super(ServerPacketType.ApplyDamage.getValue());
         writeI(sender);
         writeI(target);
         writeI(damage);
-        writeI(newHp);
+        writeI((int) newHp);
         writeB(criticalHit ? (byte) 1 : (byte) 0);
         buildPacket();
     }
