+++ gameserver/src/main/java/com/shnok/javaserver/dto/external/serverpackets/ObjectDirectionPacket.java
@@ -5,10 +5,10 @@ import com.shnok.javaserver.enums.network.packettypes.external.ServerPacketType;
 import com.shnok.javaserver.model.Point3D;
 
 public class ObjectDirectionPacket extends SendablePacket {
-    public ObjectDirectionPacket(int id, int speed, Point3D direction) {
+    public ObjectDirectionPacket(int id, float speed, Point3D direction) {
         super(ServerPacketType.ObjectMoveDirection.getValue());
         writeI(id);
-        writeI(speed);
+        writeI((int) speed);
         writeF(direction.getX());
         writeF(direction.getY());
         writeF(direction.getZ());
