+++ gameserver/src/main/java/com/shnok/javaserver/model/skills/FormulasLegacy.java
@@ -10,7 +10,7 @@ public class FormulasLegacy {
     }
 
     // Calculate delay (in milliseconds) before next ATTACK
-    public final int calcPAtkSpd(double atkSpeed) {
+    public final int calcPAtkSpd(float atkSpeed) {
         // Source L2J
         // measured Oct 2006 by Tank6585, formula by Sami
         // attack speed 312 equals 1500 ms delay... (or 300 + 40 ms delay?)
@@ -21,7 +21,7 @@ public class FormulasLegacy {
     }
 
     // Calculate delay (in milliseconds) for skills cast
-//    public final int calcMAtkSpd(Skill skill, double skillTime) {
+//    public final int calcMAtkSpd(Skill skill, float skillTime) {
 //        if (skill.isMagic()) {
 //            return (int) ((skillTime * 333) / attacker.getMAtkSpd());
 //        }
