+++ gameserver/src/main/java/com/shnok/javaserver/model/stats/functions/fomulas/FuncMAtkMod.java
@@ -18,10 +18,10 @@ public class FuncMAtkMod extends AbstractFunction {
     }
 
     @Override
-    public double calc(Entity effector, Entity effected, Skill skill, double initVal) {
+    public float calc(Entity effector, Entity effected, Skill skill, float initVal) {
         // Level Modifier^2 * INT Modifier^2
-        double lvlMod = Formulas.INTbonus[effector.getINT()];
-        double intMod = effector.isPlayer() ? effector.getLevelMod() : effector.getLevelMod();
-        return initVal * Math.pow(lvlMod, 2) * Math.pow(intMod, 2);
+        float lvlMod = Formulas.INTbonus[effector.getINT()];
+        float intMod = effector.isPlayer() ? effector.getLevelMod() : effector.getLevelMod();
+        return (float) (initVal * Math.pow(lvlMod, 2) * Math.pow(intMod, 2));
     }
 }
