+++ gameserver/src/main/java/com/shnok/javaserver/model/stats/functions/fomulas/FuncMDefMod.java
@@ -20,8 +20,8 @@ public class FuncMDefMod extends AbstractFunction {
     }
 
     @Override
-    public double calc(Entity effector, Entity effected, Skill skill, double initVal) {
-        double value = initVal;
+    public float calc(Entity effector, Entity effected, Skill skill, float initVal) {
+        float value = initVal;
         if (effector.isPlayer()) {
             PlayerInstance p = (PlayerInstance) effector;
             if (!p.getInventory().isSlotEmpty(ItemSlot.lfinger)) {
