+++ gameserver/src/main/java/com/shnok/javaserver/model/stats/functions/fomulas/FuncMaxCpMul.java
@@ -18,7 +18,7 @@ public class FuncMaxCpMul extends AbstractFunction {
     }
 
     @Override
-    public double calc(Entity effector, Entity effected, Skill skill, double initVal) {
+    public float calc(Entity effector, Entity effected, Skill skill, float initVal) {
         return initVal * Formulas.CONbonus[effector.getCON()];
     }
 }
