+++ gameserver/src/main/java/com/shnok/javaserver/model/stats/functions/fomulas/FuncMaxMpMul.java
@@ -18,7 +18,7 @@ public class FuncMaxMpMul extends AbstractFunction {
     }
 
     @Override
-    public double calc(Entity effector, Entity effected, Skill skill, double initVal) {
+    public float calc(Entity effector, Entity effected, Skill skill, float initVal) {
         return initVal * Formulas.MENbonus[effector.getMEN()];
     }
 }
